requirements.txt
paho-mqtt==1.6.1
httpx==0.23.0
python-dotenv==0.20.0
requests==2.28.0
numpy==1.23.0
scikit-learn==1.1.1
opencv-python==4.6.0.66
imutils==0.5.4
websocket-client==1.3.3
matplotlib==3.5.2
Pillow==9.1.1
PyYAML==6.0
scipy==1.8.1
torch==1.8.1
torchvision==0.9.1
tqdm==4.64.0
seaborn==0.11.2
pandas==1.4.3
thop==0.0.31.post2005241907
